Rename test_state to test_adapter and remove unneeded tests
